+++ src/python/.gitignore
@@ -48,3 +48,6 @@ pylint_*.log
 
 # tox
 .tox
+
+# Logs
+*.log
