+++ src/python/foglamp/coap/uri_handlers/sensor_values.py
@@ -7,8 +7,8 @@ import psycopg2
 import sqlalchemy as sa
 from cbor2 import loads
 from foglamp.configurator import Configurator
-from foglamp.postgres_handler import AsyncPostgresHandler
 from sqlalchemy.dialects.postgresql import JSONB
+from foglamp.foglamp_logger import *
 
 metadata = sa.MetaData()
 
@@ -19,44 +19,10 @@ __tbl__ = sa.Table(
     , sa.Column('data', JSONB))
 '''Incoming data is inserted into this table'''
 
-# Custom Sensor log levels
-DEBUG_SENSOR1 = 51
-DEBUG_SENSOR2 = 52
-
-def set_custom_log_levels():
-    """
-    This def is used to set custom log levels that may come from Coap sensor. This is
-    being done to record logs more meaningfully. Also, the custom log levels may be picked from a
-    pre defined db table.
-    """
-
-    logging.addLevelName(DEBUG_SENSOR1, "SENSOR1")
-    logging.addLevelName(DEBUG_SENSOR2, "SENSOR2")
-
-    def sensor1(self, message, *args, **kws):
-        # Yes, logger takes its '*args' as 'args'.
-        if self.isEnabledFor(DEBUG_SENSOR1):
-            self._log(DEBUG_SENSOR1, message, args, **kws)
-
-    def sensor2(self, message, *args, **kws):
-        # Yes, logger takes its '*args' as 'args'.
-        if self.isEnabledFor(DEBUG_SENSOR2):
-            self._log(DEBUG_SENSOR2, message, args, **kws)
-
-    logging.Logger.sensor1 = sensor1
-    logging.Logger.sensor2 = sensor2
-
-
 class SensorValues(resource.Resource):
     '''Handles other/sensor_values requests'''
     def __init__(self):
         super(SensorValues, self).__init__()
-
-        # Add Custom level to logging
-        set_custom_log_levels()
-
-        async_logdb = AsyncPostgresHandler()
-        logging.getLogger('coap-server').addHandler(async_logdb)
 
     def register(self, resourceRoot):
         '''Registers URI with aiocoap'''
@@ -67,6 +33,7 @@ class SensorValues(resource.Resource):
         '''Sends incoming data to database'''
         original_payload = loads(request.payload)
         
+        payload = dict(original_payload)
         payload = dict(original_payload)
 
         key = payload.get('key')
@@ -89,7 +56,7 @@ class SensorValues(resource.Resource):
                         , key # Maybe the generated key is the problem
                         , original_payload)
 
-        logging.getLogger('coap-server').log(DEBUG_SENSOR1, 'Inserted Sensor Values: %s, %s', key, payload)
+        foglamp_logger.log(DEBUG_SENSOR1, 'Inserted Sensor Values: %s, %s', key, payload)
 
         return aiocoap.Message(payload=''.encode("utf-8"))
